[solder_chk] move to a branch and deleted from main
I think this might have been doomed to not work from
the start. It's not natural for the tb to try to
drive changes that shouldn't be happening, as they
really don't want you assigning to outputs, etc.

And, I haven't fully worked out if the optimizer will
get in the way.. and finally, the ice40 has no pulldown
resistors, complicating things. Maybe I'll revisit, but
for now, at these low volumes, I'll just use a visual led
test with the pmod breakout board.