[sram] delete old version to start over
The controller was fundamentally flawed in that it
wasn't disabling we while changing data and addressees.

The test will need to be updated, and it had a couple of
wtf bugs. It will be reintroduced with tests of each part
along the way, making it clearer that it's operating
correctly.