[sram] introduce and use SB_IO for sram io
This is to get more explicit control over the timing of the
signals on the pads. The data and oe_n were not meeting the
setup and hold times in the previous version, despite seeming
to in simulation.

It is possible that this was because there was no explicit
relationship between those signals that the synthesis tools
could detect, and maybe this could have been handled in some
other way.