[vga_3to2] basic blanking
This is a fairly terrible implementation in that it just clears
pixels without any sort of fade, but it works as a POC.

nextprn can no longer find a place/route that meets timing for the
consumers of this module, e.g. adc 3to2 and gfx demo 3to2, even
with looking for seeds all the way to 256. icecube2 easily meets
timing with an fmax of 120mhz.