[gfx] disable gfx_demo_dbuf test
It was a slow test already, so it wasn't getting run frequently, but it
fails. I'm not sure if there is really a bug there or not, or if things
changed around it and it atrophied, but I don't want to sort it out now
while deep in other slow testing of the adc blanking.